apps/ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.2.3
python-dotenv==1.0.0
boto3==1.34.0
numpy>=1.26.0
pgvector==0.2.4
python-multipart==0.0.6
PyPDF2==3.0.1
Pillow>=11.0.0
pytesseract==0.3.10
python-docx==1.1.0
alembic==1.13.1
pyjwt==2.8.0
openai==1.3.7
mangum==0.17.0